Removed the svg from fragment view to fix pug bug.
Adding the text element at any position in the fragment causes the pug engine to add the points to as children of the text element rather than it's parent the svg element. Removing the text element fixes it. Moving all of the svg fragment into the parent page works fine with the text element.